project/requirements.txt
Django
djangorestframework
Pillow
stripe
python-dotenv
